fix(util): 修复k近邻计算中的数据类型问题
- 移除了torch.tensor中的dtype=torch.float64参数，避免不必要的类型转换
- 确保l2距离计算返回正确的数据类型
- 保持原有的numpy数组返回格式不变

+++ app/util.py
@@ -99,6 +99,6 @@ def k_nearest_neighbors(x, centers, k=1):
         dist (n_samples, k): k近邻距离
         indices (n_samples, k): k近邻索引
     """
-    distance = torch.tensor(l2_distance(x, centers), dtype=torch.float64)
+    distance = torch.tensor(l2_distance(x, centers))
     dist, indices = torch.topk(distance, k, dim=1, largest=False)
     return dist.numpy(), indices.numpy()
